Bring the code up to date with more recent MIT Kerberos libs.
We update the code to compile with more recent Kerberos libs.  This
involves a couple of #ifdefs.  Compile with -DOLD_MIT if your MIT
is 1.9 or less.  We define -DHAVE_KRB4 but the code is still broken,
we'll fix it if there is demand.

Also, we add a -k flag which specifies that a keytab be used rather
than passwds.